requirements.txt
langchain==0.3.2
langgraph==0.2.5
openai==1.51.0
pandas==2.2.3
python-dotenv==1.0.1